Custom Functions in Builds
I've updated the build code and Makefile. The makefile now copies over all the important files it needs to the BuildCode folder. I want to do some more work with it using Macros, but that'll be done later.

The builds now include custom functions and all the other features of the engine. The functions can be edited outside of the game and when the game is reloaded, it will reflect those changes.