Fixed Collisions and layer interactions for isGrounded
Made it so the ray cast used by isGrounded looks at the masked bits to confirm that the objects are on the same layer.
Collisions now use isTouching() to compare fixtures rather than AABBs
I also added a line in CMakeLists to the box2D pbd is properly installed, removing the build warnings when building.